A simpler collision checker
Instead of 'Math' checking the boundaries of each geometry, in this
release I've migrated to the pixel check one.

From now on this will be the default checker for any geometry allowing
granular overrides whenever needed/preffered. The square, triangle and
circle checks are still using the Math way. Then new approach can - for
now - be verified on the "arrow" element.